python_server/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
cycler==0.12.1
fonttools==4.55.3
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.0
matplotlib==3.10.0
multitasking==0.0.11
numpy==2.1.3
openpyxl==3.1.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
pymongo==4.13.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
